Extract RedshiftTestingServer from query runner
